logicForSelectedCard to save in localStorage
Без модалки, але по кліку відбувається запит по ID і записується книга в сховище масив книг

+++ src/js/home/modal-card.js
@@ -1,16 +1,16 @@
-import FetchCategoriesAll from './service-categories-all';
-
-const fetchApiCategories = new FetchCategoriesAll();
-
-export async function requestCard(bookId) {
-  const modalBook = await fetchApiCategories.getBookId(bookId);
-  console.log(modalBook);
+export async function openCardModal(selectedBook) {
   
   const bookList = JSON.parse(localStorage.getItem('SHOPPING-BOOKS-LIST'))
+
   if (bookList) {
-    bookList.push(modalBook);
-    localStorage.setItem('SHOPPING-BOOKS-LIST', JSON.stringify(bookList));
+    const bookExistsInList = bookList.find(
+      book => book._id === selectedBook._id
+    );
+    if (!bookExistsInList) {
+      bookList.push(selectedBook);
+      localStorage.setItem('SHOPPING-BOOKS-LIST', JSON.stringify(bookList));
+    }
   } else {
-    localStorage.setItem('SHOPPING-BOOKS-LIST', JSON.stringify([modalBook]));
+    localStorage.setItem('SHOPPING-BOOKS-LIST', JSON.stringify([selectedBook]));
   }
 }
